+++ CHANGELOG.md
@@ -1,6 +1,47 @@
+### Big change
+
+baked is now a npm module containing an executable!
+
+This means:
+
+- you don't have to work inside the baked.js repository anymore
+- you can use baked.js without creating any gulpfile.js and package.json file
+
+**WARNING**
+
+Because of this change, if you worked inside the baked.js directory in order to build your website, I strongly recommand to adapt you code accordingly:
+
+1. cleanup `package.json` and add `baked` in dependencies (look at `example/blank/package.json`)
+  - since this point, you can install and use the `baked` module
+2. updates your `gulpfile.js` file to load `"baked/gulp"` instead of `"./src/tasks/gulp"`
+  - since this point, you are not using the local baked.js files
+3. you can now delete the `src` directory since you don't need it anymore
+4. If you didn't do it already, you can also delete the `example` directory and the `README.md` and `CHANGELOG.md` files.
+
+### Features
+
+- Only errors are now displayed by default.
+  + (ok this is not an error but it's still cool)
+- Add `pathToHere` helper (returns the path of the current page)
+- Add `urlTo` and `urlToHere` helpers (return the full URL)
+  + do not confuse the new `urlTo` helper
+    and the old (badly named) helper `url_to` which returns only a path.
+- Add a global configuration file, which handles:
+  - the logger level
+  - the API's URL
+  - the base URL
+
+### Changes
+
+- The `url_to` helper has been renamed into `pathTo`
+  + the old version is still supported by deprecated
+- Replace `src_dir` and `dst_dir` by `srcDir` and `dstDir`, but it should not break existing gulpfiles
+  + the `init()` helper still supports `src_dir` and `dst_dir`
+  + the `parseOptions()` helper's response and the `baked.config.options` object have `src_dir` and `dst_dir` as r/w properties.
+
 ## 0.0.6
 
-###  Feature
+###  Features
 
 - Templates can now include Javascript files using `require`
   - Theses files are evaluated inside the template context
@@ -10,7 +51,7 @@
 
 ## 0.0.5
 
-### Feature
+### Features
 
 - Add template partial
   - A partial is a template whose name starts by “_”. It won't be rendered directly (so it doesn't need meta like “prismic-api”) but is included in other
@@ -29,7 +70,7 @@ twice.
 
 ## 0.0.4
 
-### Feature
+### Features
 
 - The dynamic mode is now based on the HTTP server
   - custom routes are honored in this mode too
@@ -49,7 +90,7 @@ twice.
 
 ## 0.0.3
 
-### Feature
+### Features
 
 - Add HTTP server to serve generated file
   - the generation is automatically performed when content is modified
@@ -61,7 +102,7 @@ twice.
   - the old way (`node src/server.js <src> <dst>`) is not working anymore
 - remove `.html` from generated files
 
-### Fix
+### Fixes
 
 - Fix query escape pattern (`$$` was not escaping)
 - baked: Queries can now refer to bookmarks